refactor: decompose milestones-tab into sub-components and custom hook
extract MilestoneListItem, MilestoneEditForm sub-components and
useMilestonePatch custom hook from the 629-line MilestonesTab.
move resolveValidatedEscrowData to a shared utility module.
no behavior changes.
